Add example scripts
Adding a limited set of example scripts to the "public" collection.
These scripts are not meant to be used directly.  However, they provide a
few working examples of Jenkins C.I. for an imaginary site="xyz", as a
model for others interested in implementing Jenkins C.I. within this
framework.
Also adding a table of parameter names, for reference.

Change-Id: I0cc321efd8946b7ab377a0d02901a1a1f95ae5c3
Signed-off-by: Tony Wallace <awallace@qce.qualcomm.com>