"test" jobs env inject from upstream jobs
When a "test" job is run manually (rebuild or build w parameters),
most environment variables are read as usual from the upstream job
but are not propagated to the Jenkins build environment, as they
would be if the "test" job was triggered directly by the upstream
"full" job. When that happens, any "test" job e-mail notifications
do not get the Gerrit branch. This change uses Env Inject plugin
to add the key env variables to the Jenkins build environment of
the "test" job.

Change-Id: I173eedbb699b27736fc74084430f636574f2942c
Signed-off-by: Tony Wallace <awallace@qce.qualcomm.com>